+++ packages/create-xbell/src/utils/check-is-project.ts
@@ -1,0 +1,9 @@
+import { existsSync } from 'fs';
+import { join } from 'path';
+
+export function checkIsProject(rootDir: string) {
+  const hasPkgJson = existsSync(
+    join(rootDir, 'package.json')
+  )
+  return hasPkgJson;
+}