+++ packages/create-xbell/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-xbell
 
+## 0.4.1
+
+### Patch Changes
+
+- 907f22a: fix bugs in existed project
+
+## 0.4.0
+
+### Minor Changes
+
+- bec63f8: support any test file
+
 ## 0.3.6
 
 ### Patch Changes