+++ packages/xbell/CHANGELOG.md
@@ -1,10 +1,16 @@
 # xbell
 
-## 0.4.0-alpha.0
+## 0.4.1
+
+### Patch Changes
+
+- 907f22a: fix bugs in existed project
+
+## 0.4.0
 
 ### Minor Changes
 
-- 3ebefab: browser testing
+- bec63f8: support any test file
 
 ## 0.3.6